Change the offset variable to an unsigned int since it cannot be negative and is assigned to streamOffset which is an unsigned int.
TRAC #23671

Author: Geoff Lang

+++ src/libGLESv2/renderer/IndexRangeCache.cpp
@@ -16,7 +16,7 @@
 namespace rx
 {
 
-void IndexRangeCache::addRange(GLenum type, intptr_t offset, GLsizei count, unsigned int minIdx, unsigned int maxIdx, 
+void IndexRangeCache::addRange(GLenum type, unsigned int offset, GLsizei count, unsigned int minIdx, unsigned int maxIdx,
                                unsigned int streamOffset)
 {
     mIndexRangeCache[IndexRange(type, offset, count)] = IndexBounds(minIdx, maxIdx, streamOffset);
@@ -44,7 +44,7 @@ void IndexRangeCache::invalidateRange(unsigned int offset, unsigned int size)
     }
 }
 
-bool IndexRangeCache::findRange(GLenum type, intptr_t offset, GLsizei count, unsigned int *outMinIndex,
+bool IndexRangeCache::findRange(GLenum type, unsigned int offset, GLsizei count, unsigned int *outMinIndex,
                                 unsigned int *outMaxIndex, unsigned int *outStreamOffset) const
 {
     IndexRangeMap::const_iterator i = mIndexRangeCache.find(IndexRange(type, offset, count));
